1.15.0

+++ Bam.Net.Data.Repositories.Tests/DaoRepositoryUnitTests.cs
@@ -244,7 +244,7 @@ namespace Bam.Net.Data.Repositories.Tests
         {
             DaoRepository repo = GetTestDaoRepository();
             TestContainer toCreate = new TestContainer();
-            Expect.AreEqual(0, toCreate.Id);
+            Expect.AreEqual((ulong)0, toCreate.Id);
             string testName = "Test Name".RandomLetters(5);
             toCreate.Name = testName;
             toCreate.BirthDay = DateTime.Now.Subtract(new TimeSpan(7, 0, 0, 0));
@@ -258,7 +258,7 @@ namespace Bam.Net.Data.Repositories.Tests
         {
             DaoRepository repo = GetTestDaoRepository();
             TestContainer toCreate = new TestContainer();
-            Expect.AreEqual(0, toCreate.Id);
+            Expect.AreEqual((ulong)0, toCreate.Id);
             string testName = "Test Name".RandomLetters(5);
             toCreate.Name = testName;
             toCreate.BirthDay = DateTime.Now.Subtract(new TimeSpan(7, 0, 0, 0));
